--- projects/openapi-io/src/loaders/file-on-branch.ts
@@ -1,45 +1,0 @@
-import { SpecLoaderResult } from './types';
-import {
-  parseOpenAPIFromRepoWithSourcemap,
-} from '../parser/openapi-sourcemap-parser';
-import { JsonSchemaSourcemap } from '../parser/sourcemap';
-
-const util = require('util');
-const exec = util.promisify(require('child_process').exec);
-
-export async function loadSpecFromBranch(
-  fileNameInRepo: string,
-  repoPath: string,
-  branch: string,
-  defaultValue: any,
-  includeSourcemap: boolean = true
-): Promise<SpecLoaderResult> {
-  try {
-    const results = await parseOpenAPIFromRepoWithSourcemap(
-      fileNameInRepo,
-      repoPath,
-      branch
-    );
-    return {
-      success: true,
-      flattened: results.jsonLike,
-      sourcemap: includeSourcemap
-        ? results.sourcemap
-        : new JsonSchemaSourcemap(fileNameInRepo),
-    };
-  } catch (e: any) {
-    return { success: false, error: e.message };
-  }
-}
-
-// helpers
-export async function inGit(workingDirectory: string): Promise<false | string> {
-  try {
-    const { stdout } = await exec('git rev-parse --show-toplevel', {
-      cwd: workingDirectory,
-    });
-    return stdout.trim();
-  } catch (e) {
-    return false;
-  }
-}
